Use dialog focus algorithm for dialog popovers
This patch makes it so that when a <dialog popover> gets opened as a
popover, it uses the dialog initial focus algorithm instead of the
popover initial focus algorithm.

The popover initial focus algorithm only focuses descendants with the
autofocus attribute, whereas the dialog initial focus algorithm focuses
anything thats focusable.

Fixed: 1430405
Change-Id: I627d8ab9045ecdad8d8265fab95b68e571227685