Set Opus Signal for DTX
Setting DTX for OPUS does not work, unless we explicitly set the signal
hint to `OPUS_SIGNAL_VOICE`. This workaround is already present in
WebRTC code [1], and it's unlikely that one would use DTX for music
content anyways.

This CL sets or unsets the signal hint appropriately for use with the
DTX flag.

[1]: https://chromium.googlesource.com/external/webrtc/+/bba780707848ad36066739c60d7b28cd752fb92f/webrtc/modules/audio_coding/codecs/opus/opus_interface.c#175

Bug: 1415321
Change-Id: I08ce6f392314d87fd6f7f74358fa9eb89ae75e99