Update WPTs that rely on animation timeline sibling lookup: batch2
We have spec consensus on removing sibling lookup to locate a scroll
timeline.  Before we implement this removal, we will gradually update
the tests that rely on sibling lookup.

This CL updates a non-trivial WPT where multiple siblings containing
a target's scroll-timelines are moved to form the incestry chain of
the target.

Bug: 1432627
Change-Id: I19269cb2bb84541dd88f33d5e6794a30e206d0fb