Add WPT for partitioned cookies/service worker interaction
This CL adds a WPT which exercises the interaction between partitioned service workers and partitioned cookies.

When 3P storage partitioning is enabled, partitioned cookies should
only be visible to partitioned workers whose StorageKey's equivalent CookiePartitionKey matches their own.

The CookieStoreManager API was partitioned in a separate CL.

This does NOT test that workers in nonced partitions do not have access to unpartitioned cookies.

Bug:1225444,1246549,1427879
Change-Id: Id101b4fa48bed46605fe95f7ef2b1d2463f9301a